Moving all datasource types around - this is needed to make the SDK/Document types more accesible for the development datasource plugins.
