+++ packages/server/src/integrations/couchdb.ts
@@ -2,8 +2,8 @@ import {
   Integration,
   DatasourceFieldTypes,
   QueryTypes,
-} from "../definitions/datasource"
-import { IntegrationBase } from "./base/IntegrationBase"
+  IntegrationBase,
+} from "@budibase/types"
 
 module CouchDBModule {
   const PouchDB = require("pouchdb")
