+++ packages/types/src/documents/app/row.ts
@@ -1,3 +1,23 @@
 import { Document } from "../document"
 
-export interface Row extends Document {}
+export enum FieldTypes {
+  STRING = "string",
+  LONGFORM = "longform",
+  OPTIONS = "options",
+  NUMBER = "number",
+  BOOLEAN = "boolean",
+  ARRAY = "array",
+  DATETIME = "datetime",
+  ATTACHMENT = "attachment",
+  LINK = "link",
+  FORMULA = "formula",
+  AUTO = "auto",
+  JSON = "json",
+  INTERNAL = "internal",
+}
+
+export interface Row extends Document {
+  type?: string
+  tableId?: string
+  [key: string]: any
+}
