+++ packages/types/src/sdk/index.ts
@@ -3,3 +3,5 @@ export * from "./context"
 export * from "./events"
 export * from "./licensing"
 export * from "./migrations"
+export * from "./datasources"
+export * from "./search"
